# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AssetFlow API",
    "description": "AI-native family asset allocation advisor backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/login/phone": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Login With Phone",
        "description": "Login with phone number and SMS verification code\nCreates new user if phone number doesn't exist",
        "operationId": "login_with_phone_api_v1_auth_login_phone_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PhoneLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login/device": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Login With Device",
        "description": "Anonymous login with device ID\nCreates new user if device ID doesn't exist",
        "operationId": "login_with_device_api_v1_auth_login_device_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/bind-phone": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Bind Phone To User",
        "description": "Bind phone number to current anonymous user\nRequires valid authentication token",
        "operationId": "bind_phone_to_user_api_v1_auth_bind_phone_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BindPhoneRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "authentication"
        ],
        "summary": "Get Current User Info",
        "description": "Get current user information\nRequires valid authentication token",
        "operationId": "get_current_user_info_api_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Refresh Token",
        "description": "Refresh access token\nRequires valid authentication token",
        "operationId": "refresh_token_api_v1_auth_refresh_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthResponse": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "device_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Id"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "user_id",
          "phone"
        ],
        "title": "AuthResponse",
        "description": "Authentication response"
      },
      "BindPhoneRequest": {
        "properties": {
          "phone": {
            "type": "string",
            "maxLength": 15,
            "minLength": 11,
            "title": "Phone",
            "description": "\u624b\u673a\u53f7"
          },
          "verification_code": {
            "type": "string",
            "maxLength": 6,
            "minLength": 4,
            "title": "Verification Code",
            "description": "\u9a8c\u8bc1\u7801"
          }
        },
        "type": "object",
        "required": [
          "phone",
          "verification_code"
        ],
        "title": "BindPhoneRequest",
        "description": "Bind phone to user request"
      },
      "DeviceLoginRequest": {
        "properties": {
          "device_id": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Device Id",
            "description": "\u8bbe\u5907ID"
          }
        },
        "type": "object",
        "required": [
          "device_id"
        ],
        "title": "DeviceLoginRequest",
        "description": "Device ID login request"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PhoneLoginRequest": {
        "properties": {
          "phone": {
            "type": "string",
            "maxLength": 15,
            "minLength": 11,
            "title": "Phone",
            "description": "\u624b\u673a\u53f7"
          },
          "verification_code": {
            "type": "string",
            "maxLength": 6,
            "minLength": 4,
            "title": "Verification Code",
            "description": "\u9a8c\u8bc1\u7801"
          }
        },
        "type": "object",
        "required": [
          "phone",
          "verification_code"
        ],
        "title": "PhoneLoginRequest",
        "description": "Phone login request"
      },
...
      "UserResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "device_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Id"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "phone",
          "created_at"
        ],
        "title": "UserResponse",
        "description": "User information response"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}